# Improve file read performance