… cache/cache_test.go
… stargz/fs_test.go
… stargz/reader/reader_test.go
… stargz/remote/blob.go
… stargz/remote/blob_test.go
… stargz/remote/resolver.go
… stargz/remote/util_test.go