Throughout the benchmark on #101, it turned out the file read performance is low especially on random and parallel reads. This PR includes two patches to mitigate this problem and improve file read performance.

The first patch is for squashing HTTP Range Request header on large prefetch. Current filesystem implementation chunks the prefetching range into small, many, and mostly neighbouring chunks and lists them in a single HTTP Range Request header without enough squashing. Sometimes this leads to large HTTP header and the prefetch request fails. This commit solves this by aggressively squashing neighbouring/overwrapping chunks in HTTP headers, as much as possible.

The second one contains the following optimizations for caches:
- minimizing the occurrence of byte slice allocation in the execution path of file read, leveraging sync.Pool,
- minimizing the memory copy and disk I/O by allowing to fetch a partials range of blobs from the cache, and
- minimizing the locked region in the cache.

# fio results

measured on Github Actions `ubuntu-18.04` runner.

fio conf
```
bs=4k
rw=randread
size=256M
numjobs=4
```

## master branch

See https://github.com/ktock/stargz-snapshotter/actions/runs/121204873 for more details

|image type|fio summary|memory consumption(in KB, `ps` output just after the fio completion)|
---|---|---
|legacy(ext4+overlayfs)|`READ: bw=32.6MiB/s (34.2MB/s), 8343KiB/s-8363KiB/s (8543kB/s-8564kB/s), io=1024MiB (1074MB), run=31346-31422msec`|VSZ: 873880, RSS: 32040|
|stargz|`READ: bw=1330KiB/s (1362kB/s), 333KiB/s-333KiB/s (341kB/s-341kB/s), io=1024MiB (1074MB), run=788154-788231msec`|VSZ: 3204508, RSS: 1925984|
|eStargz|`READ: bw=1571KiB/s (1609kB/s), 393KiB/s-479KiB/s (402kB/s-490kB/s), io=1024MiB (1074MB), run=547678-667481msec`|VSZ: 3813632, RSS: 1994172|

## PR branch

See https://github.com/ktock/stargz-snapshotter/actions/runs/121255820 for more details.
The bandwidth (and latency) is improved compared to the master branch.

|image type|fio summary|memory consumption(in KB, `ps` output just after the fio completion)|
---|---|---
|legacy(ext4+overlayfs)|`READ: bw=32.6MiB/s (34.2MB/s), 8345KiB/s-8353KiB/s (8545kB/s-8553kB/s), io=1024MiB (1074MB), run=31385-31414msec`|VSZ: 800168, RSS: 32316|
|stargz|`READ: bw=16.9MiB/s (17.8MB/s), 4337KiB/s-4545KiB/s (4441kB/s-4654kB/s), io=1024MiB (1074MB), run=57681-60445msec`|VSZ: 1618592, RSS: 261152|
|eStargz|`READ: bw=83.8MiB/s (87.9MB/s), 20.0MiB/s-22.2MiB/s (21.0MB/s-23.3MB/s), io=1024MiB (1074MB), run=11523-12217msec`|VSZ: 1453484, RSS: 185632|

NOTE: Stargz snapshotter waits the prefetch completion for 10s by default. But in fio+eStargz benchmark, prefetch target is large (256MB * 4 files + fio binary) so it doesn't end in 10s and the container starts before the prefetch completion. If we configure this timeout longer (e.g. 30s?) the BW should be even better.


cc: @sequix