@ktock 2020-06-01
Thanks for your comments!

> 1. Based our previsous conversation in #88,
> Won't it lead to many small files in cache directory and consume up inodes eventually?

Partial layer contents are lazily pulled so if we don't any fixed-offset for chunks it might be hard to reuse them. For example, if a process on the filesystem requires a range 3-6, there is no guarantee another process also accesses the same range. This can be 2-4 or 5-7 or whatever. Making HTTP requests every time different (but possibly overlapping) ranges is required is costly. We should pre-define fixed offset like 0-4, 5-9, ... and should download 0-9 at the first access to 3-6, which helps to avoid duplicated range request on futural 2-4 or 5-7 access. But I believe we have better ways for it. The reason why we store each chunk as a file is that this was one of the simplest ways to do this as an initial implementation of the cache.

> 2.Why do we need a ref-cnt-lru cache? From my point, there is still global lock involved in there.

I think having on-memory cache itself helps avoid slow disk I/O but I believe we might have better ways for this. This patch wraps `groupcache/lru` for making it reference-aware, keeping the locked region as minimal as possible. This is because `groupcache/lru` is agnostic about ref count of each cached objects and this can suddenly evict the oldest object (this triggers defined finalization like closing the file) even if this is being used by someone. Ideas are very welcome if you have lock-free ways for doing this.