@sequix 2020-06-02
Now, I have a total grasp of the read-path
```
                            +----------+      +-------------+
                            | registry |      |    fuse     |
                            +----+-----+      +------+------+
                                 |                   ^
                                 v                   |
                            +----+-----+      +------+--------+  fulfill fuse read requests
     merge region           | fetcher  |      | reader.Reader |  lru cache decompressed chunks
     minimize requests      +----+-----+      +------+--------+  both in memory and disk
                                 |                   ^
                                 v                   |
                            +----+-----+      +------+------+  decompress gzipped chunks
lru cache gzipped chunks    |   blob   +----->+ crfs.Reader |  provide file-level access
both in memory and disk     +----------+      +-------------+  no cache at all
```
And problems to tackle to acheive a better performance:

1. minimize network load in fetcher;
2. merge `blob` and `crfs.Reader`, or eleminate `crfs.Reader`,  in order to have a unique cache layer;
3. the unique cache should be as simple as possilbe (lockfree is the best)

 My naive thinkings:

1. fetcher should fetch fixed size chunk (better in os page size), it seems this is how things is done now;
2. migrate crfs.Reader here to get that single cache.
3. ringbuffer-based cache, it can also let the hot data in memory like lru, but with less lock granularity.

This PR lgtm, if you feel good about my thinksings above, I'll make a PR this week.