@ktock 2020-06-02
@sequix

> Now, I have a total grasp of the read-path

Great! Thanks a lot!

> - minimize network load in fetcher;
> - merge blob and crfs.Reader, or eleminate crfs.Reader, in order to have a unique cache layer;
> - the unique cache should be as simple as possilbe (lockfree is the best)

Yes, all of them are still problems.

> fetcher should fetch fixed size chunk (better in os page size), it seems this is how things is done now;

Though I don't know what is the best chunk size, fixed-size chunks are currently effective for performance.

> migrate crfs.Reader here to get that single cache.

I agree with this.

> ringbuffer-based cache, it can also let the hot data in memory like lru, but with less lock granularity.

Though we eventually should rely on benchmarking for evaluating how beneficial ring-buffer-based cache is, I think this is worth trying.

There are rooms for discussion in terms of which chunk should be stored on memory,
- Storing all accessed chunks on memory: Ideal but there possibly cause performance problem https://github.com/containerd/stargz-snapshotter/pull/105#discussion_r433222944 but we possibly can mitigate it by carefully tuning the chunk size.
- Using on-memory cache as a temporary space for chunks that are write-in-progress to disk: this PR takes this approach. It works but maybe we have better ways.